fix: split main app-frame component
